requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
PyMuPDF==1.23.26
PyPDF2==3.0.1
Pillow==10.2.0
python-dotenv==1.0.1
openai==1.12.0
reportlab==4.1.0
aiohttp==3.9.3
httpx==0.27.0
beautifulsoup4==4.12.3
Jinja2==3.1.3
pytesseract==0.3.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 